+++ assemblyadapter-list/src/androidTest/java/com/github/panpf/assemblyadapter/list/test/AssemblySingleDataListAdapterTest.kt
@@ -100,10 +100,34 @@ class AssemblySingleDataListAdapterTest {
     fun testMethodGetItemId() {
         AssemblySingleDataListAdapter(TestItemFactory()).apply {
             Assert.assertEquals(-1L, getItemId(-1))
-            Assert.assertEquals(0L, getItemId(0))
-            Assert.assertEquals(1L, getItemId(1))
-            Assert.assertEquals(Int.MAX_VALUE.toLong(), getItemId(Int.MAX_VALUE))
-            Assert.assertEquals(Int.MIN_VALUE.toLong(), getItemId(Int.MIN_VALUE))
+            Assert.assertEquals(-1L, getItemId(0))
+            Assert.assertEquals(-1L, getItemId(1))
+        }
+
+        AssemblySingleDataListAdapter(TestItemFactory(), hasStableIds = true).apply {
+            assertThrow(IndexOutOfBoundsException::class) {
+                getItemId(-1)
+            }
+            assertThrow(IndexOutOfBoundsException::class) {
+                getItemId(0)
+            }
+            assertThrow(IndexOutOfBoundsException::class) {
+                getItemId(1)
+            }
+        }
+
+        AssemblySingleDataListAdapter(
+            TestItemFactory(),
+            initData = "hello",
+            hasStableIds = true
+        ).apply {
+            assertThrow(IndexOutOfBoundsException::class) {
+                getItemId(-1)
+            }
+            Assert.assertEquals(getItem(0).hashCode().toLong(), getItemId(0))
+            assertThrow(IndexOutOfBoundsException::class) {
+                getItemId(1)
+            }
         }
     }